requirements.txt
pygithub
pyyaml